+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Base class for excluded parameters.
 - Resolving of asset/hierarchy id within tools.
+- Plot through python + plotly.
 
 ## [v0.11.4] - 4.11.2025
 
@@ -50,13 +51,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tool argument correctness metric.
 - Fix brain region get all for new hierarchy.
 - Images not appearing in Literature search (Prompt refactor).
+- Update uv lock package.
 
 ### Changed
 - Running deepeval github action also on main
 - Use typescript autogen for backend types in frontend.
 - Try to enforce using metric tools rather than downloading assets.
 - Rule to avoid overvalidating.
-- Update uv lock package.
 
 ## [v0.10.0] - 2.10.2025